document grammar and move it back to the main include file

+++ src/options.cpp
@@ -11,7 +11,7 @@ Options::Options(bool strict, std::unique_ptr<StringConverter> converter):
 {
 }
 
-Options::Options(int seed, bool strict, std::unique_ptr<StringConverter> converter):
+Options::Options(unsigned long seed, bool strict, std::unique_ptr<StringConverter> converter):
     _strict(strict),
     _rng(seed),
     _converter(std::move(converter))
